refactor: move resolver definition to crate
